Display and handle local database errors
