+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,11 @@
 <resources>
     <string name="app_name">MovieFinder</string>
-    <string name="error_no_internet">Couldn\\\'t connect to server, please check your internet connection.</string>
+    <string name="error_no_internet">Couldn\'t connect to server, please check your internet connection.</string>
     <string name="error_unknown">Something went wrong. Please try again later.</string>
-    <string name="error_serialization">Couldn\\\'t serialize or deserialize data.</string>
+    <string name="error_serialization">Couldn\'t serialize or deserialize data.</string>
     <string name="auth_failed">Authentication failed.</string>
     <string name="invalid_request">Invalid request</string>
+    <string name="db_fetch_error">Could not get data from database.</string>
+    <string name="db_upsert_error">Could not write data.</string>
+    <string name="db_delete_error">Could not delete data.</string>
 </resources>
